refactor: 필드명 변경 및 적용
