+++ api/src/main/java/io/devarium/api/controller/user/dto/UpdateUserRequest.java
@@ -5,7 +5,7 @@ import io.devarium.core.domain.user.port.UpdateUser;
 public record UpdateUserRequest(
     String name,
     String bio,
-    String picture,
+    String profileImageUrl,
     String blogUrl,
     String githubUrl
 ) implements UpdateUser {
