+++ api/src/main/java/io/devarium/api/controller/user/dto/UserResponse.java
@@ -8,7 +8,7 @@ public record UserResponse(
     String email,
     String name,
     String bio,
-    String picture,
+    String profileImageUrl,
     String blogUrl,
     String githubUrl,
     String role,
@@ -22,7 +22,7 @@ public record UserResponse(
             user.getEmail(),
             user.getName(),
             user.getBio(),
-            user.getPicture(),
+            user.getProfileImageUrl(),
             user.getBlogUrl(),
             user.getGithubUrl(),
             user.getRole().getAuthority(),
