+++ core/src/main/java/io/devarium/core/domain/user/port/OAuth2Client.java
@@ -1,6 +1,6 @@
 package io.devarium.core.domain.user.port;
 
-import io.devarium.core.domain.user.OAuth2UserInfo;
+import io.devarium.core.auth.OAuth2UserInfo;
 
 public interface OAuth2Client {
