+++ infrastructure/src/main/java/io/devarium/infrastructure/auth/oauth2/GoogleOAuth2Client.java
@@ -36,7 +36,7 @@ public class GoogleOAuth2Client implements OAuth2Client {
             userInfo.id(),
             userInfo.email(),
             userInfo.name(),
-            userInfo.picture(),
+            userInfo.profileImageUrl(),
             OAuth2Provider.GOOGLE
         );
     }
